Commit everything as-is so far, HUGE WIP
Not meant to be used by users yet, for devs only, and so I don't lose it if my hdd dies lol. Dependencies are not yet accounted for so you'll have to fiddle with them yourself if you want this to build in its current state. It needs latest HedgeLib (with the cri stuff), libRad, and some additional boost headers I added to BlueBlur in a PR (32). Should build with those.